Moved QComboBox patch to qtpy._patch
