+++ tests/test_patch_qcombobox.py
